Detector de Sellos/requirements.txt
fastapi
uvicorn
pillow
torch
opencv-python
scikit-learn
clip-by-openai
cairosvg
pdf2image 